+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
 -   Remove "overall state of detectors" line from detchar vector table,
     since without Virgo being online it will always be "bad."
 
+-   Check the iDQ OK vector before checking the value of iDQ timeseries.
+
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------